Update set command to support updating existing passwords.
This command now works by removing any matching applications (in effect
allowing for deletions), and then adding any that you specify in the
edited file.  This allows the keys to be changed (to something that
wasn't already in the file minus the matching applications), new
applications to be appended (it still fills in a template if no matching
applications found), and all the different fields to be updated for
passwords.